Adding support for Debian-12 (bookworm) -untested

+++ lib/backup.sh
@@ -13,19 +13,19 @@ mkdir -p $BACKUP_DIR
 linkables=$( find -H "$DOTFILES" -maxdepth 3 -name '*.symlink' )
 
 for file in $linkables; do
-    filename=".$( basename $file '.symlink' )"
+    filename=".$(basename "$file" '.symlink')"
     target="$HOME/$filename"
-    if [ -f $target ]; then
+    if [ -f "$target" ] && [ ! -L "$target" ]; then
         echo "backing up $filename"
-        mv $target $BACKUP_DIR
+        mv "$target" "$BACKUP_DIR"
     else
         echo -e "$filename does not exist at this location or is a symlink"
     fi
 done
 
 files=("$HOME/.config/nvim" "$HOME/.vim" "$HOME/.vimrc")
-for filename in "$HOME/.config/nvim" "$HOME/.vim" "$HOME/.vimrc"; do
-    if [ ! -L $filename ]; then
+for filename in "${files[@]}"; do
+    if [ -e "$filename" ] && [ ! -L "$filename" ]; then
         echo "backing up $filename"
         cp -rf $filename $BACKUP_DIR
     else
